+++ presentation/src/main/AndroidManifest.xml
@@ -12,6 +12,10 @@
         <activity
             android:name=".features.main.MainActivity"
             android:exported="true">
+        </activity>
+        <activity
+            android:exported="true"
+            android:name=".features.sign.signUp.SignUpActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
