+++ presentation/src/main/java/kr/hs/dgsw/smartschool/dodamdodam/features/sign/signUp/SignUpFragment1.kt
@@ -17,7 +17,7 @@ class SignUpFragment1 : BaseFragment<ItemSignUp1Binding,SignUpViewModel>(){
             textChange(edittextEmail, layoutEmail, "email")
         }
         with(viewModel) {
-            viewEvent.observe(this@SignUpActivity) { it ->
+            viewEvent.observe(this) { it ->
                 it.getContentIfNotHandled()?.let { event ->
                     when(event) {
                         kr.hs.dgsw.smartschool.dodamdodam.features.sign.signUp.SignUpViewModel.EVENT_ON_ERROR_ID -> {
@@ -41,8 +41,8 @@ class SignUpFragment1 : BaseFragment<ItemSignUp1Binding,SignUpViewModel>(){
                 state.collect { state ->
 
                     if (state.result.isNotBlank()) {
-                        android.widget.Toast.makeText(this@SignUpActivity, "회원가입 성공", android.widget.Toast.LENGTH_SHORT).show()
-                        val intent = Intent(this@SignUpActivity, SignInActivity::class.java)
+                        android.widget.Toast.makeText(this, "회원가입 성공", android.widget.Toast.LENGTH_SHORT).show()
+                        val intent = Intent(this, SignInActivity::class.java)
                         intent.putExtra("id", viewModel.id.value)
                         intent.putExtra("email", viewModel.email.value)
                         intent.putExtra("where", "signUp")
@@ -55,12 +55,33 @@ class SignUpFragment1 : BaseFragment<ItemSignUp1Binding,SignUpViewModel>(){
 
                     if (state.error.isNotBlank()) {
                         android.util.Log.d("StateLog", "SignUpActivity: ${state.error}")
-                        android.widget.Toast.makeText(this@SignUpActivity, "${state.error}", android.widget.Toast.LENGTH_SHORT).show()
+                        android.widget.Toast.makeText(this, "${state.error}", android.widget.Toast.LENGTH_SHORT).show()
                         return@collect
                     }
                 }
             }
         }
     }
+    private fun setSpinner() {
+        var yearSpinner : Spinner = mBinding.generationSpinner
 
+        ArrayAdapter.createFromResource(
+            this,
+            R.array.generation_array,
+            androidx.appcompat.R.layout.support_simple_spinner_dropdown_item
+        ).also { adapter ->
+            adapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item)
+            yearSpinner.onItemSelectedListener
+            yearSpinner.adapter = adapter
+            yearSpinner.onItemSelectedListener = object : AdapterView.OnItemSelectedListener {
+                override fun onNothingSelected(p0: AdapterView<*>?) {
+                }
+                override fun onItemSelected(p0: AdapterView<*>?, p1: View?, pos: Int, p3: Long) {
+                    //TODO(ViewModel에 값 넘겨주기)
+                    viewModel.generation.value = (pos+1).toString()
+                }
+
+            }
+        }
+    }
 }
