+++ presentation/src/main/res/drawable/bg_btn_sign_up.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android">
     <gradient
-        android:startColor="@color/main"
+        android:startColor="#CC99FF"
         android:endColor="#55BBFF"
         android:angle="0"
         android:type="linear"/>
