+++ presentation/src/main/res/layout/item_sign_up_2.xml
@@ -7,13 +7,9 @@
             name="vm"
             type="kr.hs.dgsw.smartschool.dodamdodam.features.sign.signUp.SignUpViewModel" />
     </data>
-
-        <androidx.cardview.widget.CardView
-            android:layout_width="380dp"
-            android:layout_height="400dp">
             <androidx.constraintlayout.widget.ConstraintLayout
-                android:layout_width="match_parent"
-                android:layout_height="match_parent">
+                android:layout_width="380dp"
+                android:layout_height="460dp">
         <com.google.android.material.textfield.TextInputLayout
             android:id="@+id/layout_name"
             style="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox"
@@ -51,7 +47,7 @@
                 android:weightSum="3"
                 android:background="@drawable/bg_edittext"
                 app:layout_constraintStart_toStartOf="@id/layout_name"
-                app:layout_constraintTop_toBottomOf="@id/layout_name">
+                app:layout_constraintTop_toBottomOf="@id/layout_phone">
 
                 <Spinner
                     style="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox.Dense.ExposedDropdownMenu"
@@ -98,7 +94,7 @@
                 app:helperText="휴대폰 번호를 입력하세요"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@id/layout_phone">
+                app:layout_constraintTop_toBottomOf="@id/layout_name">
 
                 <com.google.android.material.textfield.TextInputEditText
                     android:id="@+id/edittext_phone"
@@ -114,5 +110,4 @@
             </com.google.android.material.textfield.TextInputLayout>
 
         </androidx.constraintlayout.widget.ConstraintLayout>
-</androidx.cardview.widget.CardView>
 </layout>
